@@ -1,39 +1,19 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { AppComponent } from './app.component';
-import {NavbarComponent} from "./navbar/navbar.component";
-import {UserComponent} from "./user/user.component";
-import {UserService} from "./user-service/user.service";
 import {HttpClientModule} from "@angular/common/http";
-import {RouterModule} from "@angular/router";
-import { OrderComponent } from './order/order.component';
-import { ProductComponent } from './product/product.component';
-import {OrderService} from "./order-service/order.service";
-import {ProductService} from "./product-service/product.service";
-import { RegistrationComponent } from './registration/registration.component';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import {HttpClientXsrfModule} from "@angular/common/http";
+import { AppComponent } from './app.component';
+import { UserViewComponent } from './user-view/user-view.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    NavbarComponent,
-    UserComponent,
-    OrderComponent,
-    ProductComponent,
-    RegistrationComponent
+    UserViewComponent
   ],
   imports: [
-    BrowserModule, HttpClientModule,
-    HttpClientXsrfModule,
-    FormsModule,
-    ReactiveFormsModule,
-    RouterModule.forRoot([
-      { path:'offer', component: ProductComponent },
-      { path:'register', component: RegistrationComponent }
-    ])
+    BrowserModule,
+    HttpClientModule
   ],
-  providers: [UserService, OrderService, ProductService],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
