@@ -1,20 +1,23 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs";
+import {Observable} from "rxjs/index";
 import {User} from "./user";
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class UserService {
 
-  url = "/users";
-  constructor(private http: HttpClient) { }
+  url = "/users/";
 
-  public getUser(id :String): Observable<User> {
+  constructor(private http: HttpClient) {
+  }
+
+  public getUser(id: String): Observable<User> {
     return this.http.get<User>(this.url + id);
   }
 
   public createUser(user: User): Observable<User> {
     return this.http.post<User>(this.url, user);
   }
-
 }
